iox-#1142 Fix macos builds
Cast uint64_t to unsigned long to solve macos build failure

+++ iceoryx_examples/icediscovery_in_c/iox_c_find_service.c
@@ -109,7 +109,7 @@ int main()
                                                                             searchFrontCameras,
                                                                             (void*)&numberFrontCameras,
                                                                             MessagingPattern_PUB_SUB);
-        printf("\nFound %lu front cameras\n", numberFrontCameras);
+        printf("\nFound %lu front cameras\n", (unsigned long)numberFrontCameras);
 
         sleep_for(1000);
     }
